Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > (3 comments)
> > 
> > > Patch Set 8:
> > > 
> > > (3 comments)
> > > 
> > > There are definitely third-party plugins that extend GnuMakefileGenerator. This patch, in current form, may break such extenders in a severe way. I don't see an easy solution, but we could *consider* making the clean behaviour configurable (eg via a GnuMakefileGenerator.setPerFolderClean() method) if others agree. As it is, third-parties may question the validity of deprecating methods but rendering them immediately unusable.
> > 
> > The main problem (as also mentioned in the line comment) is how to decide what behavior to use. Adding a new method for it would force everyone to sub-class to enable the per directory cleaning and I assume that we want this by default to avoid the bug...?
> 
> Understood. Many/most users will see this new behaviour immediately with the current patch. But some users might have a very negative experience, their makefile generation will be broken and they will need an update from their third party extender.
> 
> If we treat this work as an *enhancement* which is enabled by the caller for the timebeing then CDT would, of course, enable the new behaviour and the majority of users would still benefit immediately. It would be the responsibility of the third party extender to update their code to take advantage of the new behaviour until the period of deprecation is concluded.
> 
> Again, I am just one voice here. It is not my intention to block this patch.

John, don't get me wrong here, but I think it's unlikely that you sub-class these 2 classes. There are so many built in assumptions about the state of the fields and in what order to call the methods to get the right result, so sub-classing and extending it would be a nightmare regardless.

@Jonah: Do you see any way to fix the bug and at the same time not upset extenders?

I might be stuck in my thoughts here, but as I see it, it comes down to what is the most important thing, and there are 2 possible ways forward:
a) Handle clean on all platforms for large projects
b) Retain old behavior and simply abandon this gerrit

For ST, there is no problem since we can simply integrate this gerrit in our copy of the makefile generator, but I would like to avoid that if possible... ;)

Patch-set: 8