Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Actually I had tested a generated project with 1500 sourcefiles giving 500MB of compile_commands.json which lead to no issues.
> So I would abandon this and wait until users really have issues.

That is really big (500MB is the file size? Must me a multiple of that in memory use in that case). I am pleased to hear that it worked with large projects. Getting the scaling right is important and too often left until an afterthought, so thanks for thinking about this now, and a relief that nothing needs to be done in the short term.

Patch-set: 3
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>