Update patch set 8
Patch Set 8:

(3 comments)

> Patch Set 8:
> 
> (3 comments)
> 
> I have given my comments on the individual methods - I think their use cases are unique.
> 
> However I think the final conclusion is to bite the bullet here and not try to maintain such backwards compatibility as the cost(effort) is too high for the benefit.
> 
> I am tempted to deprecate the whole GnuMakefileGenerator as it has terrible API and do in CDT what most CDT extenders have traditionally actually done, make a copy of GnuMakefileGenerator and apply their local changes.
> 
> That would allow extenders to keep with status quo, no breaking them, but allow us to evolve faster and be more agile.
> 
> How about we do that here. I am going to experiment with that and come back to you. If you think that is a bad idea, please let me know.

I think that sounds reasonable. I suppose we only need to copy it to an "internal" package and update all the toolchain definitions (the builder element in the extension points).

> @Torbjörn do you at ST actually extend GnuMakefileGenerator? Or do you have a copy of it like many other companies?

We currently have a copy, but I was hoping to be able to sub-class in the future. As the discussion have been going, it's likely not to happen. I suppose we (ST) should have a background task in our development process that regularly check the upstream version of the makefile generator to get the bugfixes into our copy.

Patch-set: 8