Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The only place I see a problem with this change is org.eclipse.cdt.managedbuilder.makegen.gnu.DefaultGCCDependencyCalculator3.createMakefileGenerator()
> 
> But as it is documented as " * This is an example dependency calculator that is not used by the CDT GCC tool-chain." I am not sure what to do.
> 
> Probably best to have createMakefileGenerator return a minimal interface so that different generators can all be used there without having to actually inherit from the concrete GnuMakefileGenerator

If we are going to clean up this mess, I suppose reasonable approach would be to create a small public interface that contains the two generateDummyTargets methods and then also add it as an argument to the postProcessDependencyFile method as a parameter. That way, there would be no need to instantiate another makefile generator.

But doing this would likely break more extenders... :(

I'm not sure about moving IManagedBuildGnuToolInfo into the internal package. If it's internal, I guess it could be removed as there would only be one implementation?

Patch-set: 1
CC: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>