Update patch set 8
Patch Set 8:

> Patch Set 8:
> The other alternative I see is to make the new one public and GnuMakefileGenerator2 - but lock down all the API and only open up the bits that people actually advise they are using. And make sure that the new API as it is opened up uses proper API things (like Map instead of LinkedHashMap)
> 
> I don't maintain an external one, so I defer to you and others who do on what is the correct path forward.

I think that "fixing" the API is a rather big task.
To completely tackle the issue, I think a complete redesign of how the makefile content is generated is required.
Right now, the current implementation is a hybrid of generating the content directly and collecting information to generate the makefile statements later during the process.
I think a better design would be to collect the data first and build a model of what to generate and when the model is complete, write out the make statements.
However, doing this is a huge task and I'm not sure it would be worth the effort.

Patch-set: 8