Update patch set 7
Patch Set 7:

Martin, with several concerns I highlighted already addressed, the only more important thing I see is changing the asyncExec for the Eclipse Job. Then I think it would be reasonable to go ahead and use the code of this patchset and substitute that part of the code in the existing provider. Or if you prefer a more clean-slate approach, I won't have a strong opinion either way. As long as user-facing changes are communicated in the N&N it should be OK.

Patch-set: 7
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>