Update patch set 2
Patch Set 2:

(2 comments)

This change is fine - but I don't think I understand why it is needed. Or rather, as it is needed, why don't all labels need such a change?

Perhaps it is down to the definition of "stand-alone" in the commit message? Is a "stand-alone" a label that doesn't have another control after it?

Thanks for fixing this area up, and for putting up with my questions :-)

FWIW this class is also copied from JDT and the JDT one still has a Label for that description.

Patch-set: 2
CC: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>