Update patch set 7
Patch Set 7:

(2 comments)

> Patch Set 7:
> 
> Martin, with several concerns I highlighted already addressed, the only more important thing I see is changing the asyncExec for the Eclipse Job. Then I think it would be reasonable to go ahead and use the code of this patchset and substitute that part of the code in the existing provider. Or if you prefer a 

Marc-Andre, I already changed the asyncExec as suggested but did not commit it yet. Switching the existing provider to use the mechanism in o.e.cdt.jsoncdt would require to change method processCompileCommandsFile() only. But switching it would break your use-case with clang-cl.

> more clean-slate approach, I won't have a strong opinion either way. As long as user-facing changes are communicated in the N&N it should be OK.

Yes, I prefer a more clean slate approach. But having two providers to parse the compiltation database would be confusing for CDT users. So I would prefer to abandon this gerrit and the corresponding bug and implements this in my third-party cmake-plugin.

Patch-set: 7