Update patch set 2
Patch Set 2:

> ... why don't all labels need such a change?

Only Label objects with no other context on the page (not associated with another Control) need to be explicitly converted; A label isn't labeling something else in that case, it is being used to supply some other information or meaning.  For Accessibility, trying to explain the label without context doesn't work: JAWS takes "label" at its definition and only is readable when it is tied to another control.  There are several ways to accomplish that tie, Label followed by Text or Combo is easy to identify and is done by default. For most other controls, adding via the Accessible object on the control or table and the APIS solve the issue.  Note that it is the contents that need to be tied to the control and not necessarily the Label object as the idea is to provide the information such as a title or description to the control. Generally for static label values it is easier to reuse the value as the Accessibility Event name, or description.

> Perhaps it is down to the definition of "stand-alone" in the commit message? Is a "stand-alone" a label that doesn't have another control after it?

Yes. It is a label that has no other context within the page. The usual culprits are that the Label meaning is usually to convey something at a higher level like a page description, or message, that may not directly relate to the controls.

Patch-set: 2