Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Gerald If the changes to make the control accessible are more or less standard could we try to create a new type that will take care about specified controls, like:
> ```
> AccessibleControls accessible = new AccessibleControls().accept(label).accept(button);
> ```
> ?

Most controls are already accessible, and Label is associated with the following control. 
The difficult ones tend to be a repetitive but localized fix. 
My goal was to minimize code impact in getting the desired result.
There aren't that many that controls and labels don't comply with the automatic associations.
I think the main one would be the page description, and then the label for the complex controls like tree viewers, but those all use a standard pattern. for listener either name or description.
I think a createPageDescription(String) takes care of many of these, or maybe a class DescriptionBlock more belongs in SWT side so as to not tempt use of Label.

Patch-set: 2