Update patch set 1
Patch Set 1:

> Patch Set 1: Verified-1 Code-Review-1
> 
> Build Failed 
> 
> https://ci.eclipse.org/cdt/job/cdt-verify-code-cleanliness-pipeline/3343/ : FAILURE. The code base is not clean. This means one or more of formatting, whitespace, version numbers or other checks appears to have failed. Please see the console output for details and https://github.com/eclipse-cdt/cdt-infra/tree/master/jenkins/pipelines/cdt/verify#cdt-verify-code-cleanliness for documentation on what is checked.

Key part of log:

[ERROR] Failed to execute goal org.eclipse.tycho:tycho-p2-plugin:2.5.0:p2-metadata (baselinereplace-p2-metadata) on project org.eclipse.cdt.core.native: baseline and build artifacts have same version but different contents
[ERROR]    no-classifier: different
[ERROR]       META-INF/MANIFEST.MF: different
[ERROR]          Bundle-License: not present in baseline version

It looks like perhaps Tycho 2.5.0 changed on support of deriveHeaderFromProject - see my commit message in Id536d75a92a30582418324e27d74008869f7dd3a

When I did the upgrade to 2.4 I elected to keep manifests the same - but all the bundles could get a +0.0.100 version bump and deriveHeaderFromProject removed from toplevel pom. Part of the reason was I didn't want to do the maintenance overhead of touching every bundle version, the other part was that with deriving headers like this if more header are added in future (the implication of the feature in Tycho) then we will have to do the +0.0.100 on a semi-regular basis.

I am open to ideas - including the idea that I have overcomplicated my life here!

Patch-set: 1
CC: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>