Update patch set 1
Patch Set 1:

I'm attempting to summarize whats left for this to be merged (and to add the clean fix on top of this).

Alexander mentions using x-internal. I like that idea, but dropping the .internal part of the package means that we need to use some other name. The base org.eclipse.cdt.managedbuilder.makegen.gnu is used for the old classes and cannot be marked as x-internal without causing other issues.

John raises the question about maintaining the public version on major bumps. I sort of like that, but does it make sense to put that extra work on CDT when most changes to the makefiles needs a fork anyway? Right now I am leaning towards deprecating the public classes and just "forget" about them. Does this sound reasonable?


If we decided to go this way, I only see that there needs to be added some documentation about the removal and the deprecation. It looks like the right makefile generator is called in our unit tests, so the fork appears to be working.

Anything I am missing?

Patch-set: 1