Update patch set 2
Patch Set 2:

Thank you Jonah!

I found the stray spaces after switching back to my Mac where the show spaces is working.
The first attempt to fix I tried the Gerrit in place edit and missed some. 
The area reformat was using the wrong format (working on a separate enhancement in the same workspace when I should not be.)

Patch-set: 2