Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3:
> Hopefully Anton or Tommy can comment on the amount of time, or if it should be configurable. 
> 
> 1 second seems awfully short to allow full cleanup. Any reason it shouldn't be 10seconds for example, to give maximum amount of time? The default amount of time that we wait for suspending a process recently changed from 5 to 10 seconds (and is now configurable - https://wiki.eclipse.org/CDT/User/NewIn105#Debug) so it makes sense to me that if it takes some targets up to 10 seconds to suspend, then shutting down could take that amount of time too.
> 
> Either way, that delay is a separate patch. I think the change here is good, it both cleans up the code and removes the TerminateJobObject when asking for a graceful termination.

Yes, that is a separate patch.
The needed time varies from application to application, so 10 seconds is probably too short for some, but an ocean of time for another. The time of 1 sec was introduced in commit f3745131b94ec1b7d4149b9a385d67de2c380559 back in 2002, so I suppose it's okay for most applications.
Anyway, it should not be too hard to implement since all that code is in JAVA world and should not have any impact on the native code.

Should it be a global configurable timeout or should it be something that you tell the ProcessFactory when you create the process or perhaps both?

Patch-set: 3