Bug 574131: Do not concurrently call Spawner#destroy()
As the Object#wait() voids the synchronization, more threads can call
destroy() before the Reaper thread has identified that the process
exited or the timeout occured.
The change ensures that only one call actually raises the signal while
the others are in a wait state until the process exited.

Contributed by STMicroelectronics

Change-Id: If811feef7fad35a043fe719190031eaa7fdbc6e2
Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>