Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is scary to see...

I agree 100% - as soon as I saw the title that was my reaction!

> I hope that it's okay to add an API filter for them, or is there another way to deal with constants that are expected to be constant by other components but that currently aren't?

I don't know here. I don't think it is ok because if anyone is referring to them this will cause a linkage failure. i.e. even if they are not modifying it they may be using it.

How about deprecate for deletion and make new private static finals for all of them?

WDYT?

Alternative is to make them final which may not cause linkage problems for "proper" uses.



(There is some seriously dodgy code that someone could have written with this to raise arbitrary signals though:

    Spawner p = ...
    p.NOOP = 123;
    p.isRunning()
    p.NOOP = 0;

would cause a kill(123) - but just on Linux.

I seriously wouldn't be surprised if someone did something like this to work around the lack of an arbitrary kill(int) in the Spawner API!

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>