Update patch set 1
Patch Set 1:

> Patch Set 1:
> How about deprecate for deletion and make new private static finals for all of them?
> 
> WDYT?

Sure, I was hoping that I could avoid that, but better to do it right!

> Alternative is to make them final which may not cause linkage problems for "proper" uses.

Even just final will trigger the API errors, so I suspect that it might be problematic anyway.

> I seriously wouldn't be surprised if someone did something like this to work around the lack of an 
arbitrary kill(int) in the Spawner API!

Maybe better to extend the API with that function to allow that use-case?
Spawner#raise(int, int) is available, but it's a bit ticker as you need to find the right pid (not the native pid on Windows!). Would Spawner#kill(int) be more preferable than Spawner#raise(int)?

Patch-set: 1