Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This looks ok to me. This allows consumers of the Spawner/ProcessFactory to provide a custom timeout or fallback to the global default.

So, what you mean is that there does not need to be a preference for it? I.e. you have to supply your own and no way to override it on a user basis (user of Eclipse, not someone developing a feature on top of the Spawner logic).

Patch-set: 1