Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > This looks ok to me. This allows consumers of the Spawner/ProcessFactory to provide a custom timeout or fallback to the global default.
> > > 
> > > So, what you mean is that there does not need to be a preference for it? I.e. you have to supply your own and no way to override it on a user basis (user of Eclipse, not someone developing a feature on top of the Spawner logic).
> > 
> > I was assuming you would do your TODO in the code.
> 
> Aha, I raised the question on where to put such preference in the preference system :)

Oops. I missed the toplevel comment in gerrit and focused on the code too much.

I think it is ok to have it 1 second always in the Spawner code and force integrators to add such a timeout where needed (they can hard code it, or deliver it as a preference if needed).

If you want, you can provide a java property to configure the behaviour - in the same way as other places in the native code. See static initializer of org.eclipse.cdt.utils.pty.PTY for an example

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>