Update patch set 1
Patch Set 1:

> Patch Set 1:
> Oops. I missed the toplevel comment in gerrit and focused on the code too much.
> 
> I think it is ok to have it 1 second always in the Spawner code and force integrators to add such a timeout where needed (they can hard code it, or deliver it as a preference if needed).
> 
> If you want, you can provide a java property to configure the behaviour - in the same way as other places in the native code. See static initializer of org.eclipse.cdt.utils.pty.PTY for an example

A property seams much easier than adding a preference somewhere and at the same time make it logical so users can find it if necessary.
I suppose it would be fine to do this part after M2 or should I do it right away to get it in before M2?

Patch-set: 1