Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> That seems to have done it - only those 3 tests are now failing.

I suppose so. I guess it makes little sense to keep the result in the workspace, so should we simply apply the fix in Builder.java to the tree and avoid a lot of strange errors?

Patch-set: 3
Reviewer: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>