Bug 576987 - Unable to scroll when editing a launch config
The current situation is that:
* launch configuration dialog tabs scroll correctly in the standard Eclipse
  platform dialog, but not at all in the Launchbar-specific dialog; and:
* some CDT-supplied launch configuration dialog tabs try to manage their
  own scrolling behaviour to fix scrolling in the Launchbar-specific
  dialog but this breaks scrolling for that tab in the standard Eclipse
  platform dialog.

This change fixes the launchbar-specific editor dialog to use a scrolled
composite instead of a regular composite on which to layout the content of
each tab -- This gives the launchbar's configuration editing dialog
exactly the same scrolling behaviour as the standard Eclipse platform
configuration editing dialog.

And also fixes any CDT-supplied tabs that try to manage their own
scrolling behaviour so now all tabs have the same behaviour when viewed
in the Launchbar-specific dialog as they do when viewed in the standard
Eclipse platform dialog.

Change-Id: I0d7364a24ca48bb125cae9518728b4c93b93545d
Signed-off-by: Mat Booth <mat.booth@gmail.com>