Update patch set 5
Patch Set 5:

Hi Jonah,

I think functionally this is as good as I can get it. Works on my linux computer with the known limitation that wrapping of a command line longer than two lines causes funny effects, but I have the same on gnome-terminal and vscode so I guess it's acceptable.

The feature is hard-coded (no preference to activate) but I could change it. There is probably room for performance improvements, but looks fine on my old 2-core i3.

Patch-set: 5