Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Hi Jonathan, thank you for the contribution.
> 
> Bug 528145 describes a problem with setting breakpoints from the UI. Can you explain how this patch resolves the problem please?

The existing breakpoints from the UI are never send to gdb, because the old code never call attachDebuggerToProcess. The method attachDebuggerToProcess contains the step connecting the process breakpoints to the debugger (GDBProcess_7_2).
In remote the attach is already done. The condition in the old code suppose that no attach is done. I remove condition excluding the remote session to be attach to the debugger and I modify conditions supposing that no attach are made. The attach to process is not called in case of remote session.

Patch-set: 3