Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Hi Jonathan, thank you for the contribution.
> > 
> > Bug 528145 describes a problem with setting breakpoints from the UI. Can you explain how this patch resolves the problem please?
> 
> The existing breakpoints from the UI are never send to gdb, because the old code never call attachDebuggerToProcess. The method attachDebuggerToProcess contains the step connecting the process breakpoints to the debugger (GDBProcess_7_2).
> In remote the attach is already done. The condition in the old code suppose that no attach is done. I remove condition excluding the remote session to be attach to the debugger and I modify conditions supposing that no attach are made. The attach to process is not called in case of remote session.

I think there might be a few misunderstandings here.

To my knowledge, (I can obviously be wrong here and if so, please correct me!) a remote session always involve a gdb server. It can be running on the same computer as the GDB client or on a different one that is accessible over network, serial or some other connection.
Attach to a already running process is something that is always done locally and can never happen over a network unless you also involve a GDB server, but then you are back to a "normal" remote session.


So, based on the above, I'm not sure when it's ever relevant to do the attach to process for a remote session.
@John, am I missing something here or is my understanding in-line with what the meaning of LOCAL and REMOTE in the code?

Patch-set: 3