Bug 577097: Remove requirement to have pre-integration-test phase run
The proxy server products need to be built and signed before being
moved to the correct location. Prior to this patch this happened
in pre-integration-test phase, meaning that "mvn package" would
fail to build CDT successfully.

Therefore "pull" rather than "push" proxy-server to individual bundles.
If signing is not enabled, the proxy server product won't be signed,
but that is expected.

Also, to make sure the archiving happens in package phase, we need to
have some duplication so that archiving always is listed after signing.

Change-Id: I09ef2b6384ab6f6573352f85c068756e3792512f