Update patch set 6
Patch Set 6:

Hello Jonathan,

Before you spend more time on this gerrit, can you please explain why this is needed?
I've been reading your comments several times so far and I'm still not sure in what situation it's needed.

To my understanding, we have the following use-cases in CDT:
1. Launch the application though the GDB client -> local session
2. Attach to an already running application on the same computer -> local attach session
3. Connect to a GDB server (can be both on the same computer and on another) -> remote session

Is there any other use-case that you think of?
If not, in what of the above use-cases is the breakpoints not applied to the GDB client?

To conclude, a small reproducible scenario would greatly help to understand what issue you are attempting to resolve.

Patch-set: 6