Bug 577097: Assign a category to all o.e.remote features
In the p2 site all the o.e.remote features are appearing in the
Uncategorized category, AFAICT this is because the category.xml
file was not actually valid. Note that the PTP version
in https://download.eclipse.org/tools/ptp/builds/remote/3.0/2021-09
also showed everything as uncategorized, so this commit is actually
an improvement on presentation.

Change-Id: Ic3304c8e57131efd3c7adc6eec021f72e6ba1a36