Update patch set 1
Patch Set 1: Code-Review-2

This is the change I did.
Please note that the "-l" option is too early on the command line as it's part of "${FLAGS}" for the target tool and not last on the line like "$(LIBS)" in the makefile.

I don't think these changes has any impact on the internal builder, but I've not verified it.

Patch-set: 1
CC: Gerrit User 846 <846@97ee7c02-f12f-4043-b43e-dea463d88b31>
CC: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
CC: Gerrit User 292305 <292305@97ee7c02-f12f-4043-b43e-dea463d88b31>
Reviewer: Gerrit User 178588 <178588@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2