Bug 578187: Don't discard make arguments
If a user set the make arguments manually to what the automatic
arguments would have been if enabled, they would be discarded. This
change does not do that bit of magic anymore because the user
now (since Bug 568397) edits the build command and arguments
separately so less magic needs to happen in the code.

This is a follow-up to Bug 568397.

Change-Id: Ie8f51cbb86133fc4a9506913743613712e93d881