Bug 578219: Make the FileDialog in Copy Build Log a Save dialog
Prior to this it was SWT.NONE which is now being interpreted as
an Open dialog which means an existing file must be selected.

In some conditions the SWT.NONE would allow a new file to be entered
that didn't exist already, so this did not affect everyone.

With this change a normal Save dialog is presented, and the OS confirm
overwrite is enabled.

Change-Id: Iea8c4211f8d9fe0c2f4d423780d91a62f6d3edab