Dispose of icons in managed build settings pages
In Bug 531915 the code was rewritten to use imageDescriptorFromBundle,
which lead to url being null all the time. This meant only the last
loaded image was being disposed properly.

There fix to Bug 531915 had another side effect - it changed the caching
of icons from the Image to the ImageDescriptor.

As it has been ~4 years without the image being cached, I suppose it
is ok to leave it not cached and instead I store the images to be
disposed of in a list.

Change-Id: Id3427ebfc8720da52132bd8f11714bba1e2cd0bf