Update patch set 1
Patch Set 1:

Hi Jonah, could you ok this for RC1?  The fix for Linux Tools isn't working for the
M3 CPP EPP because the CDT Docker launcher is only requiring the Docker plug-ins and the
fix was added to the p2.inf for the Docker feature.  This will have the effect of dragging in
the docker.ls editor for Dockerfiles and the docs which I believe a user should have anyway.

If you don't like this, the alternative is to copy over the p2.inf changes to the CDT Docker launcher feature so it does the same actions.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>