Bug 579666: Mark fileds and internal class for removal
This is the first step towards fixing the dark theme on windows.
In order to fix the AbstractPage class, an API breakage is required and
this will pave the way to let any sub-classers know that things are
about to change. From my point of view, the marked fields and the
InternalTab class should never have been part of the API as they are
internal to the AbstractPage class that contains proper API methods.

Contributed by STMicroelectronics

Change-Id: Icfea8c963727718d4dc99e8df916674384dcb532
Signed-off-by: Torbjörn Svensson <torbjorn.svensson@st.com>