Bug 579877 - Slow Threads update for debugging on remote target
Reason:
=======
The problem is that we use '-thread-info' command for each thread and
for each command, a lot of common conversation happen between gdb and
gdbserver for each command.

Solution:
=========
We should use '-list-thread-groups <group id>' instead which provides
the same required information for all the threads and minimizes the
communication between gdb and gdbserver. This command takes lesser time
than the collective time taken by all '-thread-info' commands.

Tested with 2 and 8 threaded application using gdb 7.1 and 11. The gain
is more prominent as the number of threads increases.

Signed-off-by: Umair Sair <umair_sair@hotmail.com>

Change-Id: Ie4824bf110f05d4136697f4e1ecbcc19dcf145a7