Bug 559318 - construct via initialisation list with single item
Add the special case handling from [dcl.list.init] so that a class
object can be constructed using an initialisation list with a single
item of the same type (or subclass type).

This is not a complete fix for all cases of initialisation lists (in
particular string literals aren't handled) but it should help for the
case identified in the bug.