Bug 580248: Allow widget baseline to be less than font baseline
The removed assertion was overly harsh and some combinations of fonts
and text meant that the assertion was being hit. This code
was based on LineNumberRulerColumn (platform.text) which has long
since removed the assertion - see
https://github.com/eclipse-platform/eclipse.platform.text/commit/3e70d44d645967b04bc2b78701c0aadc0141004b
for the removed assertion.

Note that this change simply removes the assertion and does not
address all the possibilities of mixed fonts. I suspect that there
may be some (hopefully) small drawing errors if such mixed fonts
are used. This code (unlike platform text) will continue to
assume all lines are the same height.

Change-Id: I3801b05c96e2784aabe53b3a503dbdb86e6e10c0