GitHub Action for Code Cleanliness Check
This commit runs GitHub actions for code cleanliness.
Because this includes compare and replace build,
this verifies that the commit is buildable and
clean.

The natives are not rebuilt here (yet) because
the GitHub actions runner does not have the
cross compiler tools installed.

Part of migration to GitHub - Issue #32