Bug 580505 - Fix and improve CDT.setup
Don't install org.eclipse.pde.api.tools.ee.feature.feature.group because
it's no longer in the latest update site, and that causes problems like
this one as noticed by a user:

https://www.eclipse.org/forums/index.php?t=msg&th=1111320&goto=1854139&#msg_1854139

Move all the tasks up to the project level.

Include a * requirement on the CDT targlet so that all projects are
imported by the targlet task with no need for a separate import task.
Probably many of the explicitly listed requirements are no longer needed
because all requirements of all projects in the clone are resolved now
automatically.  Also ensure that all platforms are resolved so that all
of CDT fragments are imported.

Move the Git clone task before the modular targlet task so that the
clone is available during the first resolution.

Add an EclipseIni task so that the CDT.setup is redirected into the
local clone's version such that you can modify that setup and test it
locally via Help -> Perform Setup tasks (and so that Navigate -> Open
Setup -> CDT opens the editable workspace version).

https://bugs.eclipse.org/bugs/show_bug.cgi?id=580505